refactor: UserController 리팩토링 - 프로필 관련 컨트롤러 분리
ProfileController (/profile)
- 프로필 조회, 프로필 수정, 프로필 사진 삭제, 비밀번호 변경, 이메일 변경 컨트롤러 분리
