+++ src/main/resources/templates/fragments/header.html
@@ -11,7 +11,7 @@
                    th:href="@{/user/login}">로그인</a>
             </li>
             <li class="nav-item" sec:authorize="isAuthenticated()">
-                <a class="nav-link" th:href="@{/user/settings/profile}">마이페이지</a>
+                <a class="nav-link" th:href="@{/profile}">마이페이지</a>
             </li>
             <li class="nav-item">
                 <form th:action="@{/user/logout}" th:method="post" sec:authorize="isAuthenticated()">
