+++ src/main/resources/templates/fragments/mypage/mypageTab.html
@@ -5,11 +5,11 @@
         <ul class="nav flex-column">
             <li class="nav-item" style="font-size: 20px;">
                 <a class="nav-link active" aria-current="page" href="#" style="color: #696969;"
-                   th:href="@{/user/settings/profile}">프로필</a>
+                   th:href="@{/profile}">프로필</a>
             </li>
             <li class="nav-item" style="font-size: 20px;">
                 <a class="nav-link" href="#" style="color: #696969;"
-                   th:href="@{/user/settings/profile/pwd}">비밀번호 변경</a>
+                   th:href="@{/profile/pwd}">비밀번호 변경</a>
             </li>
             <li class="nav-item" style="font-size: 20px;">
                 <a class="nav-link" href="#" style="color: #696969;"
