+++ src/main/resources/templates/user/mypage/profile.html
@@ -156,7 +156,7 @@
         function deleteImg() {
             console.log("delete img()")
             var xhr = new XMLHttpRequest();
-            xhr.open('DELETE', "/user/settings/profile/deleteimg", true);
+            xhr.open('DELETE', "/profile/img", true);
 
             xhr.onload = function () {
                 if (xhr.status === 200) {
@@ -223,7 +223,7 @@
             };
 
             let xhr = new XMLHttpRequest();
-            xhr.open("PUT", "/user/email/change", true);
+            xhr.open("PUT", "/profile/email", true);
             xhr.setRequestHeader("Content-Type", "application/json");
 
             xhr.onreadystatechange = function () {
